convex auth

+++ convex/games.ts
@@ -1,5 +1,6 @@
-import { v } from "convex/values";
+import { ConvexError, v } from "convex/values";
 import { mutation, query } from "./_generated/server";
+import { getAuthUserId } from "@convex-dev/auth/server";
 
 export const recordGame = mutation({
   args: {
@@ -8,16 +9,22 @@ export const recordGame = mutation({
     winningCombination: v.optional(v.array(v.number())),
   },
   handler: async (ctx, args) => {
+    const userId = await getAuthUserId(ctx);
+    if (userId === null) throw new ConvexError("Користувач не авторизований");
     // 1.1. Додаємо партію в таблицю games
     const gameId = await ctx.db.insert("games", {
       winner: args.winner,
       board: args.board,
       winningCombination: args.winningCombination,
       createdAt: Date.now(),
+      userId,
     });
 
     // 1.2. Оновлюємо таблицю stats в тій же транзакції
-    const currentStats = await ctx.db.query("stats").first();
+    const currentStats = await ctx.db
+      .query("stats")
+      .withIndex("by_user", (q) => q.eq("userId", userId))
+      .first();
 
     if (!currentStats) {
       await ctx.db.insert("stats", {
@@ -26,6 +33,7 @@ export const recordGame = mutation({
         winsO: args.winner === "O" ? 1 : 0,
         draws: args.winner === "DRAW" ? 1 : 0,
         updatedAt: Date.now(),
+        userId,
       });
     } else {
       await ctx.db.patch(currentStats._id, {
@@ -47,9 +55,12 @@ export const recordGame = mutation({
 export const getRecentGames = query({
   args: {},
   handler: async (ctx) => {
+    const userId = await getAuthUserId(ctx);
+    if (userId === null) return [];
+
     return await ctx.db
       .query("games")
-      .withIndex("by_creation")
+      .withIndex("by_user", (q) => q.eq("userId", userId))
       .order("desc")
       .take(20);
   },
@@ -61,6 +72,13 @@ export const deleteGame = mutation({
     id: v.id("games"),
   },
   handler: async (ctx, args) => {
+    const userId = await getAuthUserId(ctx);
+    if (userId === null) throw new ConvexError("Користувач не авторизований");
+    const game = await ctx.db.get(args.id);
+    if (!game) throw new ConvexError("Гру не знайдено");
+    if (game.userId != userId)
+      throw new ConvexError("Нема прав на видалення цього запису");
+
     await ctx.db.delete(args.id);
     return { success: true };
   },
@@ -70,7 +88,13 @@ export const deleteGame = mutation({
 export const clearAllGames = mutation({
   args: {},
   handler: async (ctx) => {
-    const allGames = await ctx.db.query("games").collect();
+    const userId = await getAuthUserId(ctx);
+    if (userId === null) throw new ConvexError("Користувач не авторизований");
+
+    const allGames = await ctx.db
+      .query("games")
+      .withIndex("by_user", (q) => q.eq("userId", userId))
+      .collect();
     for (const game of allGames) {
       await ctx.db.delete(game._id);
     }
